05-AI Multi Agent Chatbot/multi-agent-app/requirements.txt
openai
langchain
pandas
langchain_experimental
langchain_openai
tabulate
azure-storage-blob